Refactor get friends method
