+++ src/controllers/users_controllers/self_user_controller.js
@@ -1,12 +1,15 @@
 const express = require('express');
-const {getUserById,
+const {friendsRouter} = require("./friends_controller");
+const {getUserById} = require('../../services/users_services/users_getter_service');
+const {
     deleteUserById,
-    updateUserById,
-} = require('../../services/users_service');
+    updateUserById
+} = require('../../services/users_services/users_editing_service');
 const {asyncWrapper} = require('../../utils/async_wrapper');
 
 // eslint-disable-next-line new-cap
 const router = express.Router();
+router.use('/friends', friendsRouter);
 
 router.get('/', asyncWrapper(async (req, res) => {
     const user = await getUserById(req.user._id);
